// libs/reports/statistic/common/src/lib/new-report-statistics-route.enum.ts
export enum NewReportStatisticsRouteEnum {
    Custom = 'general',
    Statistics = 'statistics',
    Clicks = 'clicks',
    Conversions = 'conversions',
    InvalidTraffic = 'invalid-traffic',
    InvalidClicks = 'invalid-clicks',
    AdvertiserPostbacks = 'advertiser-postbacks',
    AffiliatesPostbacks = 'affiliates-postbacks',
    Adjustments = 'adjustments',
    Referrals = 'referrals',
    SmartLinks = 'smart-links',
    Day = 'day',
    Month = 'month',
    Affiliate = 'affiliate',
    Advertiser = 'advertiser',
    Offer = 'offer',
    Goal = 'goal',
    Country = 'country',
    DeviceType = 'device-type',
    ConnectionType = 'connection-type',
    Os = 'os'
}