// libs/reports/statistic/common/src/lib/models/breakdown.enum.ts
export enum BreakdownEnum {
    Custom = 'general',
    Affiliate = 'affiliate',
    Offer = 'offer',
    Advertiser = 'advertiser',
    Country = 'country',
    Goals = 'goal',
    Day = 'day',
    Hour = 'hour',
    Month = 'month',
    Year = 'year',
    ConnectionType = 'connection_type',
    DeviceType = 'device_type',
    DeviceOS = 'device_os',
    DeviceBrand = 'device_brand',
    DeviceModel = 'device_model',
    DeviceOSVersion = 'devices_os_version',
    AffiliateSource = 'source',
    AffiliateSubId1 = 'sub_id1',
    AffiliateSubId2 = 'sub_id2',
    AffiliateSubId3 = 'sub_id3',
    AffiliateSubId4 = 'sub_id4',
    AffiliateSubId5 = 'sub_id5',
    AffiliateParam1 = 'aff_param1',
    AffiliateParam2 = 'aff_param2',
    AffiliateParam3 = 'aff_param3',
    AffiliateParam4 = 'aff_param4',
    AffiliateParam5 = 'aff_param5',
    ClickRefererUrl = 'click_referer',
    DeepLinkUrl = 'deep_link',
    GoalType = 'goal_type',
    Currency = 'currency',
    Creative = 'creative',
    LandingPage = 'link',
    AdvertiserTrackId = 'track_id',
    AdvertiserParam1 = 'adv_param1',
    AdvertiserParam2 = 'adv_param2',
    AdvertiserParam3 = 'adv_param3',
    AdvertiserParam4 = 'adv_param4',
    AdvertiserParam5 = 'adv_param5',
    ConversionRefererUrl = 'conversion_referer',
    Browser = 'browser',
    Language = 'language',
    MobileOperator = 'mobile_operator',
    SmartLink = 'smartlink',
    PaidToAffiliate = 'paid_to_affiliate',
    AffiliateInvoice = 'affiliate_invoice',
    AdvertiserManager = 'advertiser_manager',
    AffiliateManager = 'affiliate_manager'
}